requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
bcrypt==3.1.7
black==19.10b0
cffi==1.14.0
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-MySQL==1.5.1
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql==0.0.2
mysqlclient==1.4.6
pathspec==0.7.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
regex==2020.2.20
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
